+++ NEWS.md
@@ -1,6 +1,7 @@
 # SIBER 2.1.2.9000
 + Contains a hotfix owing to a change to 'spatstat' detailed below
 + swapped package `spatstat` for `spatstat.utils` as per instructions from their package maintainers 23/03/2017
++ please install via `devtools::install_github("andrewljackson/SIBER", build_vingettes = TRUE)`
 
 # SIBER 2.1.2
 + added ability to specify custom pch point orders in `plotSiberObject()`